@@ -1,4 +1,7 @@
 /**
- * 시스템 전반에서 사용되는 상수 정의를 포함하는 패키지
+ * 상수 정의를 포함하는 패키지
+ * 
+ * 주요 클래스:
+ * - Constants: 공통, 시스템, 에러 관련 상수 정의
  */
 package io.csh.utils.core.constants; 
