@@ -1,4 +1,9 @@
 /**
- * CSH Utils에서 사용되는 예외 클래스들을 포함하는 패키지
+ * 예외 처리 관련 클래스들을 포함하는 패키지
+ * 
+ * 주요 클래스:
+ * - CshException: 기본 예외 클래스
+ * - CshRuntimeException: 런타임 예외 클래스
+ * - ErrorCode: 에러 코드 정의
  */
 package io.csh.utils.core.exception; 
