@@ -1,4 +1,11 @@
 /**
- * CSH Utils의 핵심 기능을 제공하는 패키지
+ * Core 모듈은 모든 모듈에서 공통으로 사용되는 기본 기능을 제공합니다.
+ * 
+ * 주요 기능:
+ * - 설정 관리 (config 패키지)
+ * - 상수 정의 (constants 패키지)
+ * - 예외 처리 (exception 패키지)
+ * - 유틸리티 (util 패키지)
  */
+package io.csh.utils.core;
  
