@@ -1,4 +1,9 @@
 /**
- * 기본 유틸리티 클래스들을 포함하는 패키지
+ * 유틸리티 클래스들을 포함하는 패키지
+ * 
+ * 주요 클래스:
+ * - FileUtil: 파일 관련 유틸리티
+ * - CompareUtil: 비교 관련 유틸리티
+ * - StringUtil: 문자열 관련 유틸리티
  */
 package io.csh.utils.core.util; 
