@@ -1,12 +1,24 @@
 package io.csh.utils.logging;
 
-/**
- * 로그 레벨을 정의하는 enum
- */
 public enum LogLevel {
-    TRACE,  // 가장 상세한 로그 레벨
-    DEBUG,  // 디버깅을 위한 상세 정보
-    INFO,   // 일반적인 정보
-    WARN,   // 경고
-    ERROR   // 오류
+    TRACE(0),
+    DEBUG(1),
+    INFO(2),
+    WARN(3),
+    ERROR(4),
+    FATAL(5);
+
+    private final int level;
+
+    LogLevel(int level) {
+        this.level = level;
+    }
+
+    public int getLevel() {
+        return level;
+    }
+
+    public boolean isEnabled(LogLevel other) {
+        return this.level <= other.level;
+    }
 } 
